pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loom-agent"
version = "0.1.0"
description = "MCP server for extracting video frames from Loom URLs"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "pydantic>=2.0.0",
    "yt-dlp>=2025.12.8",
    "openai-whisper>=20231117",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/loom_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
